refactor: move services to integration package and update imports
