+++ src/main/kotlin/fi/elsapalvelu/elsa/service/integration/OkHttpClientBuilder.kt
@@ -1,4 +1,4 @@
-package fi.elsapalvelu.elsa.service
+package fi.elsapalvelu.elsa.service.integration
 
 import okhttp3.OkHttpClient
