+++ src/main/kotlin/fi/elsapalvelu/elsa/service/integration/OpintosuorituksetFetchingService.kt
@@ -1,4 +1,4 @@
-package fi.elsapalvelu.elsa.service
+package fi.elsapalvelu.elsa.service.integration
 
 import fi.elsapalvelu.elsa.domain.enumeration.YliopistoEnum
 import fi.elsapalvelu.elsa.service.dto.OpintosuorituksetPersistenceDTO
