+++ src/main/kotlin/fi/elsapalvelu/elsa/service/integration/peppi/uef/PeppiUefOpintotietodataFetchingServiceImpl.kt
@@ -3,8 +3,8 @@ package fi.elsapalvelu.elsa.service.integration.peppi.uef
 import fi.elsapalvelu.elsa.config.ApplicationProperties
 import fi.elsapalvelu.elsa.domain.enumeration.YliopistoEnum
 import fi.elsapalvelu.elsa.repository.YliopistoRepository
-import fi.elsapalvelu.elsa.service.OkHttpClientBuilder
-import fi.elsapalvelu.elsa.service.OpintotietodataFetchingService
+import fi.elsapalvelu.elsa.service.integration.OkHttpClientBuilder
+import fi.elsapalvelu.elsa.service.integration.OpintotietodataFetchingService
 import fi.elsapalvelu.elsa.service.dto.OpintotietodataDTO
 import fi.elsapalvelu.elsa.service.integration.peppi.PeppiCommonOpintotietodataFetchingServiceImpl
 import org.springframework.beans.factory.annotation.Qualifier
