+++ src/main/kotlin/fi/elsapalvelu/elsa/service/integration/sisu/tampere/SisuTreAuthenticationTokenServiceImpl.kt
@@ -8,7 +8,7 @@ import fi.elsapalvelu.elsa.config.ApplicationProperties
 import fi.elsapalvelu.elsa.security.AuthenticationToken
 import fi.elsapalvelu.elsa.security.AuthenticationTokenCache
 import fi.elsapalvelu.elsa.service.AuthenticationTokenService
-import fi.elsapalvelu.elsa.service.OkHttpClientBuilder
+import fi.elsapalvelu.elsa.service.integration.OkHttpClientBuilder
 import fi.elsapalvelu.elsa.service.constants.JSON_DATA_PROSESSING_ERROR
 import fi.elsapalvelu.elsa.service.constants.JSON_FETCHING_ERROR
 import fi.elsapalvelu.elsa.service.constants.JSON_MAPPING_ERROR
