+++ src/main/kotlin/fi/elsapalvelu/elsa/service/integration/sisu/tampere/SisuTreClientBuilderImpl.kt
@@ -4,7 +4,7 @@ import fi.elsapalvelu.elsa.config.ApplicationProperties
 import fi.elsapalvelu.elsa.interceptor.OkHttp3RequestInterceptor
 import fi.elsapalvelu.elsa.security.AccessTokenAuthenticator
 import fi.elsapalvelu.elsa.service.AuthenticationTokenService
-import fi.elsapalvelu.elsa.service.OkHttpClientBuilder
+import fi.elsapalvelu.elsa.service.integration.OkHttpClientBuilder
 import okhttp3.OkHttpClient
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.stereotype.Service
