+++ src/test/kotlin/fi/elsapalvelu/elsa/externalintegration/sisu/hy/SisuHyExternalIntegrationTests.kt
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.databind.DeserializationFeature
 import fi.elsapalvelu.elsa.config.ApplicationProperties
 import fi.elsapalvelu.elsa.externalintegration.FetchingServiceExternalIntegrationBase
 import fi.elsapalvelu.elsa.repository.YliopistoRepository
-import fi.elsapalvelu.elsa.service.OpintotietodataFetchingService
-import fi.elsapalvelu.elsa.service.OpintosuorituksetFetchingService
+import fi.elsapalvelu.elsa.service.integration.OpintotietodataFetchingService
+import fi.elsapalvelu.elsa.service.integration.OpintosuorituksetFetchingService
 import fi.elsapalvelu.elsa.service.integration.sisu.SisuTutkintoohjelmaFetchingService
 import fi.elsapalvelu.elsa.service.integration.sisu.helsinki.SisuHyClientBuilderImpl
 import fi.elsapalvelu.elsa.service.integration.sisu.helsinki.SisuHyOpintosuorituksetFetchingServiceImpl
